remove sections, align center
