+++ app/changelog/changelog.tsx
@@ -1,0 +1,42 @@
+const updates = [
+  // {
+  //   date: "99 Month Year",
+  //   title: "Updates 🌠🌠🌠",
+  //   description: (
+  //     <>
+  //       <p>describe changes here.</p>
+  //       <p></p>
+  //       <p></p>
+  //     </>
+  //   ),
+  // },
+  // {
+  //   date: "6th September 2024",
+  //   title: "🌟 v.0.0.2 - Claude Sonnet Model, UI/UX improvements.",
+  //   description: (
+  //     <>
+  //       <ul className="text-gray-60 list-disc space-y-1 ps-6 text-sm">
+  //         <li>New AI Model - Clude 3.5 Sonnet</li>
+  //         <li>GPT Autocompletions</li>
+  //         <li>UI/UX improvements</li>
+  //         <li>Bug Fixes like some extensions not working.</li>
+  //       </ul>
+  //     </>
+  //   ),
+  // },
+  {
+    date: "15 July 2024",
+    title: "🚀 Initial Launch v0.0.1",
+    description: (
+      <>
+        <ul className="text-gray-60 list-disc space-y-1 ps-6 text-sm">
+          <li>VSCode Fork - feel right at home</li>
+          <li>Chat with AI models who have full code context</li>
+          <li>tag your files @filename</li>
+        </ul>
+      </>
+    ),
+  },
+];
+
+export default updates;
